ODIN-22 Rename model.ts classes
