+++ src/app/models/AlertType.model.ts
