+++ src/app/models/SceneSettings.model.ts
