+++ src/app/models/logs/MeshLog.model.ts
