+++ src/app/models/menuActions/SidebarPanelAction.model.ts
