+++ src/app/models/panels/GeometryPanels.model.ts
