+++ src/app/services/index.service.ts
@@ -1,5 +1,5 @@
 export { AppService } from './app.service';
-export { CanvasHelperService } from './canvas-helper.service';
+export { CanvasService } from './canvas.service';
 export { EngineService } from './engine.service';
 export { LibraryService } from './library.service';
 export { LogService } from './log.service';
