+++ src/app/shared/alert/alert.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { AlertType } from './../../models/AlertType';
+import { AlertType } from '../../models/AlertType.model';
 
 @Component({
   selector: 'alert',
