+++ src/app/store/reducers/session.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import { login } from '../actions';
-import { User } from './../../models/User';
+import { User } from '../../models/User.model';
 
 export interface State {
     user: User;
