+++ src/app/ui/material-panel/material-panel.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { select, Store } from '@ngrx/store';
 import { Color3, Mesh, StandardMaterial } from 'babylonjs';
 import { filter, map } from 'rxjs/operators';
-import { AppState } from '../../store/reducers/app.reducer';
+import { AppState } from '../../store/app.reducer';
 import { EngineService, LogService } from './../../services/index.service';
 
 @Component({
