+++ src/app/ui/scene-settings/scene-settings.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { PopupDialogAction } from 'src/app/models/actions/PopupDialogAction';
+import { PopupDialogAction, SceneSettings } from '../../models';
 import { openSceneSettings } from '../../store/actions';
-import { AppState } from '../../store/reducers/app.reducer';
-import { SceneSettings } from './../../models/SceneSettings';
+import { AppState } from '../../store/app.reducer';
 import { AppService, EngineService, LogService } from './../../services/index.service';
 import { BtnFooter } from './../../shared/popup-window/popup-window.component';
