+++ src/app/ui/text-panel/text-panel.component.ts
@@ -3,7 +3,7 @@ import { select, Store } from '@ngrx/store';
 import { Mesh } from 'babylonjs';
 import { filter, map } from 'rxjs/operators';
 import { Container } from 'src/app/engine/common/Container';
-import { AppState } from '../../store/reducers/app.reducer';
+import { AppState } from '../../store/app.reducer';
 import { EngineService, LogService } from './../../services/index.service';
 
 @Component({
